% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataProc.R
\name{breakBin}
\alias{breakBin}
\title{featureSelection
Subset the genes, given a matrix}
\usage{
breakBin(x, breakVec)
}
\arguments{
\item{x}{One gene expression sample profile.}

\item{breakVec}{a vector of probabilities}
}
\value{
xbin, binned expression profile
}
\description{
featureSelection
Subset the genes, given a matrix
}
\examples{
Xsub <- featureSelection(Xmat, Ybin, 0.1)

}
