% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callSubtypes.R
\name{callEnsemble}
\alias{callEnsemble}
\title{callEnsemble
Make subtype calls for each sample}
\usage{
callEnsemble(ens, X, Y)
}
\arguments{
\item{ens}{list, result of fitEnsembleModel}

\item{X}{gene expression matrix, genes in rows, samples in columns}
}
\value{
table, column 1 is best call, remaining columns are subtype prediction scores.
}
\description{
callEnsemble
Make subtype calls for each sample
}
\examples{
calls <- callEnsemble(mods, X, Y)

}
