% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callSubtypes.R
\name{callEnsemble2}
\alias{callEnsemble2}
\title{callEnsemble2
Make subtype calls for each sample}
\usage{
callEnsemble2(X, path = "data", geneids = "symbol", datasource = "RNA-seq")
}
\arguments{
\item{X}{gene expression matrix, genes in row.names, samples in column.names}

\item{path}{the path to the ensemble model, stored as RData, and named 'ens'}

\item{geneids}{either hgnc for gene symbols or entrez ids. will be matched to the EB++ matrix}

\item{datasource}{either RNA-seq or io360 for the Nanostring platform}
}
\value{
table, column 1 is best call, remaining columns are subtype prediction scores.
}
\description{
callEnsemble2
Make subtype calls for each sample
}
\examples{
calls <- callEnsemble(mods, X, Y)

}
