% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callSubtypes.R
\name{callOneSubtype}
\alias{callOneSubtype}
\title{callOneSubtype
Make subtype calls for one sample}
\usage{
callOneSubtype(mods, X, ci)
}
\arguments{
\item{mods}{xgboost model list}

\item{X}{gene expression matrix, genes in rows, samples in columns}

\item{ci}{cluster label, and index into mods}
}
\value{
preds of one cluster model.
}
\description{
callOneSubtype
Make subtype calls for one sample
}
\examples{
calli <- callOneSubtype(mods, X, 4)

}
