% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callSubtypes.R
\name{callSubtypes}
\alias{callSubtypes}
\title{callSubtypes
Make subtype calls for each sample}
\usage{
callSubtypes(mods, X, cores, dtype, mtype)
}
\arguments{
\item{mods}{xgboost model list}

\item{X}{gene expression matrix, genes in rows, samples in columns}

\item{cores}{num cores -- not used, but should be used.}

\item{dtype}{data type}
}
\value{
table, column 1 is best call, remaining columns are subtype prediction scores.
}
\description{
callSubtypes
Make subtype calls for each sample
}
\examples{
calls <- callSubtypes(mods, X)

}
