% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataProc.R
\name{createPairsFeatures}
\alias{createPairsFeatures}
\title{createPairsFeatures
given a matrix and gene pairs, create binary features}
\usage{
createPairsFeatures(X, genes)
}
\arguments{
\item{X}{One gene expression matrix}

\item{genePairs}{a vector of gene pairs}
}
\value{
xbin, binned expression profile
}
\description{
createPairsFeatures
given a matrix and gene pairs, create binary features
}
\examples{
Xsub <- createPairsFeatures(Xmat, genepairs)

}
