% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitModels.R
\name{cvFitOneModel}
\alias{cvFitOneModel}
\title{cvfitOneModel
Train a single subtype model using cross validation}
\usage{
cvFitOneModel(Xbin, Ybin, params = list(max_depth = 2, eta = 0.5, nrounds
  = 100, nthread = 5, nfold = 5))
}
\arguments{
\item{Xbin}{Gene expression matrix.}

\item{Ybin}{Phenotype vector.}
}
\value{
A single xgboost classifier.
}
\description{
cvfitOneModel
Train a single subtype model using cross validation
}
\examples{
modC1 <- fitOneModel(ebppGeneExpr, phenotype)

}
