% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataProc.R
\name{dataProc}
\alias{dataProc}
\title{dataProc
Data preprocessing}
\usage{
dataProc(X, mods = NULL, ci = NA)
}
\arguments{
\item{mods}{a model or list of models, containing breakpoints, used to bin expression data}

\item{ci}{the cluster label and index into list of models}

\item{Xmat}{Matrix of gene expression, genes in columns, samples in rows}
}
\value{
Xbin, the binned, subset, and binarized values.
}
\description{
dataProc
Data preprocessing
}
\examples{
mod1 <- dataProc(X, mods)

}
