% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainRobencla.R
\name{editPairList}
\alias{editPairList}
\title{Edit a pair_list, remove pairs containing identified items.}
\usage{
editPairList(pair_list, items_to_remove)
}
\arguments{
\item{pair_list}{Vector or named list of vectors. Pairs ordered as (1,2), (3,4), (5,6), etc.}

\item{items_to_remove}{Vector of items; pairs containing any of these will be removed.}
}
\value{
new_pair_list Same structure as input (vector or list) with pairs removed.
}
\description{
Return the cleaned pair_list.
}
\examples{
\dontrun{
  new_list <- editPairList(pair_list, 'IGJ')
}

}
\keyword{internal}
