% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataProc.R
\name{extractGenes}
\alias{extractGenes}
\title{extractGenes
Extract the genes needed for the models.}
\usage{
extractGenes(mods)
}
\arguments{
\item{mods}{The trained xboost models, list or single.}
}
\value{
genes A vector of gene identifiers
}
\description{
extractGenes
Extract the genes needed for the models.
}
\examples{
genes <- extractGenes(mods)
}
