% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataProc.R
\name{featureSelection}
\alias{featureSelection}
\title{featureSelection
Subset the genes, given a matrix}
\usage{
featureSelection(Xmat, Ybin, testRes, ptail = 0.05)
}
\arguments{
\item{Xmat}{Matrix of gene expression data.}

\item{Ybin}{Binary phenotype vector.}
}
\value{
Xsub, subset of Xmat by genes
}
\description{
featureSelection
Subset the genes, given a matrix
}
\examples{
Xsub <- featureSelection(Xmat, Ybin, 0.1)

}
