% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitModels.R
\name{fitAllModels}
\alias{fitAllModels}
\title{fitAllModels
Training all subtype models.}
\usage{
fitAllModels(x, y)
}
\arguments{
\item{x}{Gene expression matrix.}

\item{y}{Phenotype vector.}
}
\value{
A list containing xgboost classifiers.
}
\description{
fitAllModels
Training all subtype models.
}
\examples{
mod1 <- fitAllModels(ebppGeneExpr, phenotype)

}
