% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitModels.R
\name{fitEnsembleModel}
\alias{fitEnsembleModel}
\title{fitEnsembleModel
Train a single subtype model using cross validation}
\usage{
fitEnsembleModel(Xs, Ys, n = 5, sampSize = 0.6, breakVec = c(0, 0.25,
  0.5, 0.75, 1), params = list(max_depth = 2, eta = 0.5, nrounds = 100,
  nthread = 5, nfold = 5), ptail = 0.05)
}
\arguments{
\item{Xs}{Gene expression matrix.}

\item{Ys}{Phenotype vector, multiclass}

\item{n}{Size of the ensember, where each member is a result from fitSubtypeModel}

\item{sampSize}{proportion of samples to hold back}

\item{params}{Parameters for xgboost}
}
\value{
A list of lists of xgboost classifiers
}
\description{
fitEnsembleModel
Train a single subtype model using cross validation
}
\examples{
mods <- fitSubtypeModel(Xs, Ys, params)

}
