% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callSubtypes.R
\name{geneMatch}
\alias{geneMatch}
\title{geneMatch
Match the incoming data to what was used in training}
\usage{
geneMatch(X, geneid = "pairs")
}
\arguments{
\item{X}{gene expression matrix, genes in rows, samples in columns}
}
\value{
X, but with genes matching the ebpp set, missing genes are NAs
}
\description{
geneMatch
Match the incoming data to what was used in training
}
\examples{
Xprime <- geneMatch(X)

}
