% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callSubtypes.R
\name{geneMatch2}
\alias{geneMatch2}
\title{geneMatch2
Match the incoming data to what was used in training}
\usage{
geneMatch2(X, geneid = "pairs", mode)

geneMatch2(X, geneid = "pairs", mode)
}
\arguments{
\item{X}{gene expression matrix, genes in rows, samples in columns}
}
\value{
X, but with genes matching the ebpp set, missing genes are NAs

X, but with genes matching the ebpp set, missing genes are NAs
}
\description{
geneMatch2
Match the incoming data to what was used in training

geneMatch2
Match the incoming data to what was used in training
}
\examples{
Xprime <- geneMatch(X)

Xprime <- geneMatch(X)

}
