% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainRobencla.R
\name{getFeaturesGeneTable}
\alias{getFeaturesGeneTable}
\title{Default Gene List for Immune Subtype Classification}
\usage{
getFeaturesGeneTable()
}
\value{
A named list with the 36 used genes.
}
\description{
Returns the default named genes used for immune subtype classification.
}
\examples{
gene_list <- getFeaturesGeneTable()

}
