% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelPerf.R
\name{importantFeatures}
\alias{importantFeatures}
\title{Build the importance table.}
\usage{
importantFeatures(mods, mtype = "ensemble", ci = 1)
}
\arguments{
\item{mods}{list of models fit to each subtype}

\item{type}{the type of models, single, subtypes(list of 6), ensemble}
}
\value{
table of informative features
}
\description{
Build the importance table.
}
\examples{
mod1 <- getImportance(mods, type='ensemble')

}
