% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelPerf.R
\name{modelPerf}
\alias{modelPerf}
\title{modelPerf
Checking model performance.}
\usage{
modelPerf(bst, Xbin, Ybin, title = "perf1")
}
\arguments{
\item{bst}{A trained xgboost}

\item{Xbin}{Binned gene expression data}

\item{Ybin}{Binary phenotype vector.}
}
\value{
list of err, the error in prediction, and a rocPlot
}
\description{
modelPerf
Checking model performance.
}
\examples{
mod1 <- fitAllModels(ebppGeneExpr, phenotype)

}
