% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelPerf.R
\name{modelPerf2}
\alias{modelPerf2}
\title{Checking model performance.}
\usage{
modelPerf2(calls, Ytest, subtype = 1)
}
\arguments{
\item{calls}{Calls from callSubtypes(.)}

\item{Ytest}{Multi-class phenotype vector.}

\item{subtype}{The subtype label}
}
\value{
list of err, the error in prediction, and a rocPlot
}
\description{
Checking model performance.
}
\examples{
mod1 <- fitAllModels(ebppGeneExpr, phenotype)

}
