% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callSubtypes.R
\name{reportError}
\alias{reportError}
\title{Print Gene Match Error Report}
\usage{
reportError(err)
}
\arguments{
\item{err}{Numeric, the match error proportion (0-1).}
}
\value{
NULL (called for side effect of printing).
}
\description{
Display a formatted report of the gene matching error rate.
}
\keyword{internal}
