% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelPerf.R
\name{subtypePerf}
\alias{subtypePerf}
\title{Checking model performance.}
\usage{
subtypePerf(calls, Ytest)
}
\arguments{
\item{Ytest}{Multi-class phenotype vector.}

\item{mods}{list of models fit to each subtype}
}
\value{
list of lists of err, the error in prediction, and a rocPlot
}
\description{
Checking model performance.
}
\examples{
mod1 <- fitAllModels(ebppGeneExpr, phenotype)

}
