% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataProc.R
\name{testBinFun}
\alias{testBinFun}
\title{testBinFun
Get difference in sums}
\usage{
testBinFun(G, Ybin)
}
\arguments{
\item{G}{Binary values around a previously selected pivot point}

\item{Ybin}{Binary phenotype vector.}
}
\value{
test result, numeric value, the rank sum test.
}
\description{
testBinFun
Get difference in sums
}
\examples{
res1 <- testBinFun(G, Ybin)
}
