% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataProc.R
\name{testFun}
\alias{testFun}
\title{testFun
Get difference in mean rank sums for a single gene}
\usage{
testFun(rankg, Ybin)
}
\arguments{
\item{rankg}{Gene expression profile for single sample}

\item{Ybin}{Binary phenotype vector.}
}
\value{
test result, numeric value, the rank sum test.
}
\description{
testFun
Get difference in mean rank sums for a single gene
}
\examples{
res1 <- testFun(G, Ybin)
}
