% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataProc.R
\name{trainDataProc}
\alias{trainDataProc}
\title{trainDataProc
Data preprocessing}
\usage{
trainDataProc(Xmat, Yvec, testRes = NULL, cores = 2, cluster = "1",
  tail = 0.05, breakVec = c(0, 0.25, 0.5, 0.75, 1))
}
\arguments{
\item{Xmat}{Matrix of gene expression, samples in columns, genes in rows}

\item{Yvec}{Vector of phenotype, strings, 1, 2, etc}

\item{testRes}{List of previously calculated test results, from testFun(.)}

\item{cluster}{cluster name, string '1', as in Yvec}

\item{tail}{Binary phenotype vector.}

\item{breakVec}{vector of break points, used to bin expression data}
}
\value{
List of Xbin and Ybin, the binned, subset, and binarized values.
}
\description{
trainDataProc
Data preprocessing
}
\examples{
mod1 <- trainDataProc(Xmat, Yvec, tail, cluster, breakVec==c(0, 0.25, 0.5, 0.75, 0.85, 1.0))

}
