% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callSubtypes.R
\name{callEnsemble}
\alias{callEnsemble}
\title{callEnsemble
Make subtype calls for each sample}
\usage{
callEnsemble(X, path = "data", geneids = "symbol")
}
\arguments{
\item{X}{gene expression matrix, genes in row.names, samples in column.names}

\item{path}{the path to the ensemble model, stored as RData, and named 'ens'}

\item{geneids}{either hgnc for gene symbols or entrez ids. will be matched to the EB++ matrix}
}
\value{
table, column 1 is best call, remaining columns are subtype prediction scores.
}
\description{
callEnsemble
Make subtype calls for each sample
}
\examples{
calls <- callEnsemble(mods, X, Y)

}
