% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitModels.R
\name{cvFitOneModel}
\alias{cvFitOneModel}
\title{cvfitOneModel
Train a single subtype model using cross validation}
\usage{
cvFitOneModel(Xbin, Ybin, params = list(max_depth = 2, eta = 0.5, nrounds
  = 100, nthread = 5, nfold = 5), breakVec = c(0, 0.25, 0.5, 0.75, 1),
  genes)
}
\arguments{
\item{Xbin}{Binned and filtered gene expression matrix.}

\item{Ybin}{Binned phenotype vector.}
}
\value{
A single xgboost classifier.
}
\description{
cvfitOneModel
Train a single subtype model using cross validation
}
\examples{
res0 <- trainDataProc(Xmat, Y, cluster='1')
dat  <- res0$dat
modC1 <- fitOneModel(dat$Xbin, dat$Ybin)

}
