% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callSubtypes.R
\name{geneMatchErrorReport}
\alias{geneMatchErrorReport}
\title{geneMatchErrorReport
Check whether the incoming data matches the 485 model gene IDs}
\usage{
geneMatchErrorReport(X, geneid = "pairs")
}
\arguments{
\item{X}{gene expression matrix, genes in rows, samples in columns}
}
\value{
list with percent missing genes and a vector of missing genes
}
\description{
geneMatchErrorReport
Check whether the incoming data matches the 485 model gene IDs
}
\examples{
missingGenes <- geneMatchErrorReport(X)

}
