% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelPerf.R
\name{modelPerf2}
\alias{modelPerf2}
\title{Checking model performance when Y is multiple valued}
\usage{
modelPerf2(calls, Ytest, subtype = NA)
}
\arguments{
\item{calls}{Calls from callSubtypes(.)}

\item{Ytest}{Multi-class phenotype vector.}

\item{subtype}{The subtype label}
}
\value{
list of err, the error in prediction, and a rocPlot
}
\description{
Checking model performance when Y is multiple valued
}
\examples{
mod1 <- fitAllModels(ebppGeneExpr, phenotype)

}
